6_deployment/torched-fruit/requirements.txt
alabaster==0.7.12
Babel==2.6.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.7
Flask-WTF==0.14.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud==0.34.0
google-cloud-core==0.29.1
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.0
numpy==1.15.4
numpydoc==0.8.0
olefile==0.46
packaging==18.0
Pillow==4.2.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.3.1
pyparsing==2.3.0
pytz==2018.7
requests==2.21.0
rsa==4.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
torch==1.0.0
torchvision==0.2.1
urllib3==1.24.2
Werkzeug==0.14.1
WTForms==2.2.
